1. Added console_log to gitignore 2. Added todo to Dockerfile 3. Removed messaging from remove_empty_csvs to eliminate unneccesary messaging to console_log 4. Added a FATAL switchy to kill the backtest if the constituent list for a particular month is empty. 5. Added sink() to 0_all.R to send all console output to a file 6. Added dialog to 0_all.R to ask if it should delete data directory (if re-downloading) and if it should delete trials and results directories 7. Minor string formatting improvement to print statement in 1_query_source.R 8. Moved price_related_data and volume_data variables to parameters.R, since these change depending on data source 9. Fixed naming in bloomberg-to_datalog so that there is consistency in ticker names between constituent lists, metadata files and marketdata and fundamentaldata filenames 10. created a market_identifier and fundamental_identifier field in metadata files in bloomberg_to_datalog so that donwstream scripts know which fields to parse to lookup ticker data 11. Fixed load_slow_moving_data.R so that it can take any price and volume related fields (as specified in parameters) and create a price matrix. Also fills NA with last for price data and NA with 0 for volume data to prevent NA errors. 12. Minor fix to random_trial.R to geenrate random values for any number of constituents. Previoously it was hardcoded to 50. 13. Deleted all the other random trials. Will add back at some point but didn't want the hardcoded sample floating about and too lazy to change them all manually.

+++ scripts/0_all.R
@@ -1,3 +1,4 @@
+sink("console_log")
 # Install required packages
 list.of.packages <- c("Rblpapi", 
                       "here", 
@@ -10,21 +11,51 @@ new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"
 if(length(new.packages)) install.packages(new.packages)
 
 # Reset all directories
-unlink("data", recursive=T )
-unlink("trials", recursive=T)
+delete_data <- readline(prompt=paste("Delete all datasets and datalog? [N/y]: ", sep=" "))
+if(delete_data == "y") {
+  print("Deleting all datasets...")
+  unlink("data", recursive=T )
+} else {
+  print("Not deleting all datasets and datalog...")
+}
+
+# Delete trials
+delete_trials <- readline(prompt=paste("Delete all files in the trials directory? [N/y]: ", sep=" "))
+if(delete_trials == "y") {
+  print("Deleting all files in the trials directory...")
+  unlink("trials", recursive=T)
+  # Copy in the sample trials
+  dir.create("trials")
+  print("Copying in samples to the trials directory...")
+  file.copy(from=file.path("scripts/trading/sample_trials",list.files("scripts/trading/sample_trials")), 
+            to="trials", 
+            overwrite = TRUE, recursive = FALSE, 
+            copy.mode = TRUE)
+  
+} else {
+  print("Not deleting all files in the trials directory...")
+}
+
+print("Deleting any tempfiles...")
 unlink("temp", recursive=T)
+print("Deleting any old results...")
 unlink("results", recursive=T)
 
-# Copy in the sample trials
-dir.create("trials")
-file.copy(from=file.path("scripts/trading/sample_trials",list.files("scripts/trading/sample_trials")), 
-          to="trials", 
-          overwrite = TRUE, recursive = FALSE, 
-          copy.mode = TRUE)
-
 # Run the entire processing chain
+print("Running 1_query_source.R")
 source("scripts/1_query_source.R")
+
+print("Running 2_process_data.R")
 source("scripts/2_process_data.R")
+
+print("Running 3_run_trials.R")
 source("scripts/3_run_trials.R")
+
+print("Running 4_report.R")
 source("scripts/4_report.R")
+
+print("Running 5_cross_validate.R")
 source("scripts/5_cross_validate.R")
+
+sink(type="message")
+sink()
